+++ plugin_tooling/src-tests/com/github/rustdt/tooling/ops/RacerOutputParserTest.java
@@ -16,9 +16,12 @@ import static melnorme.utilbox.core.CoreUtil.listFrom;
 import java.util.List;
 
 import melnorme.lang.tests.CommonToolingTest;
-import melnorme.lang.tooling._actual.ToolCompletionProposal;
+import melnorme.lang.tooling.ToolCompletionProposal;
+import melnorme.lang.tooling.ToolingMessages;
 import melnorme.lang.tooling.completion.LangCompletionResult;
+import melnorme.lang.tooling.ops.FindDefinitionResult;
 import melnorme.lang.tooling.ops.OperationSoftFailure;
+import melnorme.lang.tooling.ops.SourceLineColumnRange;
 import melnorme.utilbox.core.CommonException;
 
 import org.junit.Test;
@@ -31,7 +34,7 @@ public class RacerOutputParserTest extends CommonToolingTest {
 	public void test$() throws Exception {
 		int offset = 10;
 		
-		RacerOutputParser buildParser = new RacerOutputParser(offset) {
+		RacerCompletionOutputParser buildParser = new RacerCompletionOutputParser(offset) {
 			@Override
 			protected void handleLineParseError(CommonException ce) {
 				assertFail();
@@ -61,10 +64,39 @@ public class RacerOutputParserTest extends CommonToolingTest {
 		);
 	}
 	
-	protected void testParseOutput(RacerOutputParser parser, String output, List<?> expected) 
+	protected void testParseOutput(RacerCompletionOutputParser parser, String output, List<?> expected) 
 			throws CommonException, OperationSoftFailure {
 		LangCompletionResult result = parser.parse(output);
 		assertAreEqualLists((List<?>) result.getValidatedProposals(), expected);
+	}
+	
+	@Test
+	public void parseMatchLine() throws Exception { parseMatchLine$(); }
+	public void parseMatchLine$() throws Exception {
+		
+		RacerCompletionOutputParser buildParser = new RacerCompletionOutputParser(0) {
+			@Override
+			protected void handleLineParseError(CommonException ce) {
+				assertFail();
+			}
+		};
+		
+		testParseResolvedMatch(buildParser, "", 
+			new FindDefinitionResult(ToolingMessages.FIND_DEFINITION_NoTargetFound));
+		
+		testParseResolvedMatch(buildParser, 
+			"MATCH other,19,3,/devel/RustProj/src/main.rs,Function,fn other(foo: i32) {", 
+			new FindDefinitionResult(null, new SourceLineColumnRange(path("/devel/RustProj/src/main.rs"), 19, 4)));
+		
+		testParseResolvedMatch(buildParser, "MATCH other,as", 
+			new FindDefinitionResult("Error: Invalid integer: `as`"));
+		
+	}
+	
+	protected void testParseResolvedMatch(RacerCompletionOutputParser buildParser, String input,
+			FindDefinitionResult expected) {
+		FindDefinitionResult result = buildParser.parseResolvedMatch(input);
+		assertAreEqual(result, expected);
 	}
 	
 }